+++ packages/label-core/src/barrelExports.test.ts
@@ -45,6 +45,7 @@ import * as usFoodNetQuantitySeparationRule from './rules/usFood/netQuantitySepa
 import * as usFoodNetQuantityTypeSizeRule from './rules/usFood/netQuantityTypeSize'
 import * as usFoodResponsibleFirmRule from './rules/usFood/responsibleFirm'
 import * as usFoodDualColumnRule from './rules/usFood/dualColumn'
+import * as usFoodDualColumnFormRule from './rules/usFood/dualColumnForm'
 import * as usFoodStatementOfIdentityRule from './rules/usFood/statementOfIdentity'
 import { GHS_RULES, GS1_RETAIL_RULES, US_FOOD_RULES } from './rules/registry'
 import * as findingBuilders from './rules/finding'
@@ -195,6 +196,7 @@ const MODULES: ReadonlyArray<readonly [name: string, barrel: object, members: Me
       ['usFood/informationPanelTypeSize.ts', usFoodInformationPanelTypeSizeRule],
       ['usFood/responsibleFirm.ts', usFoodResponsibleFirmRule],
       ['usFood/dualColumn.ts', usFoodDualColumnRule],
+      ['usFood/dualColumnForm.ts', usFoodDualColumnFormRule],
       ['usFood/statementOfIdentity.ts', usFoodStatementOfIdentityRule],
     ],
   ],