+++ packages/label-core/src/rules/registry.ts
@@ -36,6 +36,7 @@ import {
   usFoodNutritionRoundingRule,
 } from './usFood/nutritionFacts'
 import { usFoodDualColumnRule } from './usFood/dualColumn'
+import { usFoodDualColumnFormRule } from './usFood/dualColumnForm'
 import { usFoodNutritionFormatRule } from './usFood/nutritionFormat'
 import { usFoodNutritionTypeSizeRule } from './usFood/nutritionTypeSize'
 import { usFoodNetQuantityDualDeclarationRule } from './usFood/netQuantityDualDeclaration'
@@ -131,6 +132,7 @@ export const US_FOOD_RULES: readonly UsFoodRule[] = [
   usFoodNutritionTypeSizeRule,
   usFoodNutritionFormatRule,
   usFoodDualColumnRule,
+  usFoodDualColumnFormRule,
   usFoodResponsibleFirmRule,
   usFoodInformationPanelTypeSizeRule,
 ]